move page sections to constants
