+++ apps/web/src/hooks/useRoutes.ts
@@ -3,8 +3,8 @@ import { useIntl } from "react-intl";
 
 import { Locales, getLocale } from "@gc-digital-talent/i18n";
 
-import { PageSectionId as UserProfilePageSectionId } from "~/components/UserProfile/constants";
-import { PageSectionId as CareerTimelineAndRecruitmentPageSectionId } from "~/pages/Profile/CareerTimelineAndRecruitmentPage/constants";
+import { PageSectionId as UserProfilePageSectionId } from "~/constants/sections/userProfile";
+import { PageSectionId as CareerTimelineAndRecruitmentPageSectionId } from "~/constants/sections/careerTimeline";
 
 export const FromIapDraftQueryKey = "fromIapDraft";
 export const FromIapSuccessQueryKey = "fromIapSuccess";
