+++ apps/web/src/pages/ProfileAndApplicationsPage/components/TrackApplications/TrackApplications.tsx
@@ -9,7 +9,7 @@ import { FragmentType, getFragment, graphql } from "@gc-digital-talent/graphql";
 
 import useRoutes from "~/hooks/useRoutes";
 import { isApplicationInProgress } from "~/utils/applicationUtils";
-import { PAGE_SECTION_ID as CAREER_TIMELINE_AND_RECRUITMENTS_PAGE_SECTION_ID } from "~/pages/Profile/CareerTimelineAndRecruitmentPage/constants";
+import { PAGE_SECTION_ID as CAREER_TIMELINE_AND_RECRUITMENTS_PAGE_SECTION_ID } from "~/constants/sections/careerTimeline";
 import ApplicationCard from "~/components/ApplicationCard/ApplicationCard";
 
 function buildLink(href: string, chunks: ReactNode): ReactElement {
